refactor: make lifecycle the canonical owner of repo/YAML utilities
Move repo resolution, YAML I/O, configured-versions, and pg-lifecycle
into the rhdh_lifecycle package. Prow accesses them via a symlink and
thin re-export facades in rhdh_prow, requiring zero import changes in
prow consumer scripts.
