+++ skills/prow/scripts/rhdh_lifecycle
@@ -1,0 +1,1 @@
+../../lifecycle/scripts/rhdh_lifecycle
